Created tests for the Subaccounts endpoints
